Fix bugs and add features to DrProf
getSlowestPass and getFastestPass should not initialize with the
duration of the first time stamp if the latter is incomplete. This is
now fixed.

getPasses now properly counts only complete passes. A new feature,
getIncompletePasses counts only the incomplete passes.

When profiling the View, we now profile the time it takes Qt to swap the
frame. Apparently, this creates quite a bottleneck!